Faction Progress saved
Refactored the action costs into factions and faction costs. Things now only are dependant on one tower type at most and the rules reflect this. Also added a struct that tracks that most towers of each faction you've had on the board. This value is now the limiting factor to what actions you can take.